Added js to make calculator work
Storing the first number to calculate in firstValue
variable. Then after the math symbol is clicked,
the symbol will also be stored in symbol variable.
Then, the second number will be stored in
secondValue variable. Finally after the equal button
is clicked, the operate function will be called and
passed three parameters: firstValue, secondValue, and
symbol.